+++ src/Http/Controllers/Inertia/RemoveConnectedAccountsController.php
@@ -7,38 +7,9 @@ use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\ValidationException;
-use Laravel\Jetstream\Http\AuthenticatesConnectedAccounts;
-use Laravel\Socialite\Facades\Socialite;
 
-class ConnectedAccountsController extends Controller
+class RemoveConnectedAccountsController extends Controller
 {
-    use AuthenticatesConnectedAccounts;
-
-    /**
-     * Get the redirect for the given Socialite provider.
-     *
-     * @param  string  $provider
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    public function show(string $provider)
-    {
-        return Socialite::driver($provider)->redirect();
-    }
-
-    /**
-     * Attempt to log the user in via the provider user returned from Socialite.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  string  $provider
-     * @return \Illuminate\Routing\Pipeline
-     */
-    public function store(Request $request, string $provider)
-    {
-        return $this->authenticate(
-            $request, $provider, Socialite::driver($provider)->user()
-        );
-    }
-
     /**
      * Delete a connected account.
      *
