--- src/Http/Controllers/Livewire/ConnectedAccountsController.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Laravel\Jetstream\Http\Controllers\Livewire;
-
-use Illuminate\Http\Request;
-use Illuminate\Routing\Controller;
-
-class ConnectedAccountsController extends Controller
-{
-    public function show(Request $request, string $provider)
-    {
-
-    }
-
-    public function store(Request $request, string $provider)
-    {
-
-    }
-}
